Separate the HTTP method from the path in the route.
A route is the addition of a path and an HTTP method.  The HTTP method
is stored along the path for each route.

By default, if a route is created without the HTTP method specified,
routes for the GET PUT POST DELETE methods are created.

This should allow some introspection on route in the future.