Add default OPTIONS route.
Unless this feature is disabled, a default OPTIONS route is added for
every exiting route.  Also, add a 405 Method Not Available for all the
non existing methods for a given route.

This way, if you define a GET and DELETE for a route, but no POST, PUT
and PATCH, a 405 will be returned to the client if it tries to reach one
of this endpoint.